Update research.txt
I updated the cost section that I had to finish, and I fixed the Geothermal Section- Bc I made a mistake by not adding the "What Is It" part. So add that to the website as well. Also, go to the ProjectGreenLight google doc to find the table for the cost section. it should be shared with all of you. if it is not ttm I will fix that.